backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
opengradient
python-dotenv>=1.0.0
pydantic>=2.10.0
feedparser>=6.0.0
